Added 'boundary' to the 'Content-Type' param in the headers during a request
I kept receiving a IterationStop error when parsing my JSON, and the problem was there was no boundary. Not sure if others are having this problem but I thought I'd upload it just in case. By the way, I really like your adaptation of supersimplemediawiki! It has been really helpful.